+++ Eyes.Selenium.DotNet/Fluent/ISeleniumCheckTarget.cs
@@ -14,6 +14,9 @@ namespace Applitools.Selenium.Fluent
         CheckState State { get; set; }
 
         List<int> GetLayoutBreakpoints();
+
         bool GetLayoutBreakpointsEnabled();
+
+        bool? GetUseCookies();
     }
 }